requirements.txt
Flask>=2.0.0
Flask-SQLAlchemy>=2.5.0
Flask-Migrate>=3.0.0
SQLAlchemy>=1.4.0
Flask-RQ2>=18.0
rq==2.3.1
Flask-CORS>=3.0.0
PyJWT>=2.0.0
Flask-JWT-Extended>=4.0.0 # Added Flask-JWT-Extended
Flask-Limiter>=3.0.0
python-dotenv>=0.19.0
gunicorn>=20.0.0
setuptools>=68.0.0,<81.0.0

# RCON Service dependencies
pyzmq>=25.0.0
redis>=5.0.0

# Flask-SocketIO for WebSocket support
Flask-SocketIO>=5.3.0
python-socketio>=5.0.0
eventlet>=0.35.0
python-json-logger>=2.0.0
requests>=2.31.0
